Fix feature flag errors
