+++ crates/relayer-cli/src/commands.rs
@@ -16,14 +16,11 @@ mod update;
 mod upgrade;
 mod version;
 
-#[cfg(feature = "experimental")]
-mod new_relay;
-
 use self::{
     clear::ClearCmds, completions::CompletionsCmd, config::ConfigCmd, create::CreateCmds,
     fee::FeeCmd, health::HealthCheckCmd, keys::KeysCmd, listen::ListenCmd,
-    misbehaviour::MisbehaviourCmd, new_relay::NewRelayCmds, query::QueryCmd, start::StartCmd,
-    tx::TxCmd, update::UpdateCmds, upgrade::UpgradeCmds, version::VersionCmd,
+    misbehaviour::MisbehaviourCmd, query::QueryCmd, start::StartCmd, tx::TxCmd, update::UpdateCmds,
+    upgrade::UpgradeCmds, version::VersionCmd,
 };
 
 use core::time::Duration;
@@ -103,7 +100,8 @@ pub enum CliCmd {
 
     /// Relay packets using the new experimental relayer architecture.
     #[clap(subcommand)]
-    Relay(NewRelayCmds),
+    #[cfg(feature = "experimental")]
+    Relay(crate::next::commands::new_relay::NewRelayCmds),
 }
 
 /// This trait allows you to define how application configuration is loaded.
