+++ crates/relayer-cli/src/next/commands/mod.rs
@@ -1,0 +1,1 @@
+pub mod new_relay;
